shorten org settings component names, go to kebab case on all
