--- src/app/modules/organizations/organizations-settings/organizations-settings.component.html
@@ -1,5 +1,0 @@
-<div class="prose">
-  @if (organization$ | async; as o) {
-    Organization: {{ o.name }}
-  }
-</div>
