+++ src/app/modules/organizations/settings/audit-template/audit-template.component.ts
@@ -5,11 +5,11 @@ import { OrganizationService } from '@seed/api/organization'
 import { SharedImports } from '@seed/directives'
 
 @Component({
-  selector: 'seed-organizations-settings-api-key',
-  templateUrl: './api_key.component.html',
+  selector: 'seed-organizations-settings-audit-template',
+  templateUrl: './audit-template.component.html',
   imports: [CommonModule, SharedImports, MatIconModule],
 })
-export class APIKeyComponent {
+export class AuditTemplateComponent {
   private _organizationService = inject(OrganizationService)
   organization$ = this._organizationService.currentOrganization$
 }
