+++ src/app/modules/organizations/settings/default-display-fields/default-display-fields.component.ts
@@ -5,11 +5,11 @@ import { OrganizationService } from '@seed/api/organization'
 import { SharedImports } from '@seed/directives'
 
 @Component({
-  selector: 'seed-organizations-settings-two-factor',
-  templateUrl: './two_factor.component.html',
+  selector: 'seed-organizations-settings-default-display-fields',
+  templateUrl: './default-display-fields.component.html',
   imports: [CommonModule, SharedImports, MatIconModule],
 })
-export class TwoFactorComponent {
+export class DefaultDisplayFieldComponent {
   private _organizationService = inject(OrganizationService)
   organization$ = this._organizationService.currentOrganization$
 }
