--- src/app/modules/organizations/settings/default_display_fields/default_display_fields.component.ts
@@ -1,15 +1,0 @@
-import { CommonModule } from '@angular/common'
-import { Component, inject } from '@angular/core'
-import { MatIconModule } from '@angular/material/icon'
-import { OrganizationService } from '@seed/api/organization'
-import { SharedImports } from '@seed/directives'
-
-@Component({
-  selector: 'seed-organizations-settings-default-display-fields',
-  templateUrl: './default_display_fields.component.html',
-  imports: [CommonModule, SharedImports, MatIconModule],
-})
-export class DefaultDisplayFieldComponent {
-  private _organizationService = inject(OrganizationService)
-  organization$ = this._organizationService.currentOrganization$
-}
