+++ src/app/modules/organizations/settings/settings.component.ts
@@ -20,20 +20,20 @@ export class SettingsComponent {
       type: 'basic',
     },
     {
-      id: 'organizations/settings/api_key',
-      link: '/organizations/settings/api_key',
+      id: 'organizations/settings/api-key',
+      link: '/organizations/settings/api-key',
       title: 'API Key',
       type: 'basic',
     },
     {
-      id: 'organizations/settings/audit_template',
-      link: '/organizations/settings/audit_template',
+      id: 'organizations/settings/audit-template',
+      link: '/organizations/settings/audit-template',
       title: 'Audit Template',
       type: 'basic',
     },
     {
-      id: 'organizations/settings/default_display_fields',
-      link: '/organizations/settings/default_display_fields',
+      id: 'organizations/settings/default-display-fields',
+      link: '/organizations/settings/default-display-fields',
       title: 'Default Display Fields',
       type: 'basic',
     },
@@ -56,8 +56,8 @@ export class SettingsComponent {
       type: 'basic',
     },
     {
-      id: 'organizations/settings/two_factor',
-      link: '/organizations/settings/two_factor',
+      id: 'organizations/settings/two-factor',
+      link: '/organizations/settings/two-factor',
       title: 'Two Factor Authentication',
       type: 'basic',
     },
