--- src/app/modules/organizations/settings/two_factor/two_factor.component.html
@@ -1,11 +1,0 @@
-<h3 class="text-2xl font-bold">Require Two-Factor Authentication</h3>
-<div class="prose">
-  Checking this box requires all organization users to use Two-Factor Authentication (2FA). If a user does not already have a 2FA method configured, an email-based authentication method will be set as the default.
-
-
-</div>
-<div class="prose">
-  @if (organization$ | async; as o) {
-    Organization API Key: {{ o.name }}
-  }
-</div>
